+++ .eslintrc.js
@@ -11,7 +11,7 @@ module.exports = {
     'json',
   ],
   parserOptions: {
-    ecmaVersion: 2017,
+    ecmaVersion: 2018,
   },
   ignorePatterns: [
     '!.eslintrc.js',
